Fix undefined instance variable @parent_type warning
When running the test suite, the following warning appears *many* times:

```
warning: instance variable @parent_type not initialized
```

We have to wrap both in conditionals because if you review `symbols_for_association_chain` carefully it may not set the instance variable leaving it in a still undefined state. In case it is, we fall on the default behavior of returning nil. This is passes all the tests while removing all warnings for this variable.